requirements.txt
torch
torchvision
timm
numpy
Pillow
streamlit
